frontend/requirements.txt
streamlit
python-dotenv
bs4
